+++ src/window/mod.rs
@@ -8,6 +8,7 @@ use crate::sys;
 use crate::{Condition, Ui};
 
 pub(crate) mod content_region;
+pub(crate) mod scroll;
 
 bitflags! {
     /// Window hover check option flags